Task 4.3 without floats

+++ Lab4_ADC/Main Lab Task/Main Lab Task/adc.h
@@ -5,6 +5,6 @@
 
 void adc_init();
 uint16_t adc_read(uint8_t channel);
-float adc_convert_mv(uint16_t value);
+uint16_t adc_convert_mv(uint16_t value);
 
 #endif /* ADC_H_ */
